Moves Correlations to provisioned Tempo data source.
This negates the need to create a Tempo data source
after the stack has come up, and also removes the need
to use a separate service to populate the Correlations.

Also ensure appl services always get restarted, and
fixed an issue with the OTel Collector now that the
Service Graph processor has become a connector.

Signed-off-by: Heds Simons <hedley.simons@grafana.com>