Updates configuration for local and Grafana Cloud execution.
The `ctl.sh` scripts and unconfigured YAML for Grafana Cloud was
slightly cumbersome. With the advent of Flow and the ability to
read from local files, this commit allows the Agent to simply use
a common JSON object located in file to configure the URL and
credentials for each signal.

The README includes details on how to configure for Grafana Cloud,
which is now simply a matter of replacing the URLs and basic auth
credentials in the `agent/endspoints-cloud.json` file.

Also a slight modification to the endpoints code to standardise
naming conventions.

Finally, static configurations for the Agent have now been
deprecated and the relevant files and options removed.

Signed-off-by: Heds Simons <hedley.simons@grafana.com>

Update

Signed-off-by: Heds Simons <hedley.simons@grafana.com>