Add Grafana Cloud Frontend functionality to the Frontend app.
This will allow the use of a Frontend endpoint to be used,
to use Grafana Frontend functionality in Cloud.

The `mythical-server` service has been updated to provide configuration
for the Faro Web SDK to ensure that either a local instance of
Alloy is used for data collection, or Grafana Cloud Frontend.

Signed-off-by: Heds Simons <hedley.simons@grafana.com>